network: rename project name
